requirements.txt
imagehash
Pillow
moviepy
librosa
ffmpeg